[#103]refactor: element_manager의 call_id, depth를 관리하는 기능을 visualization_manager.py 생성해서 분리
